+++ server/auth-service/src/main/java/com/project/authservice/service/RoleService.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface RoleService {
     List<RoleDto> getAllRoles();
-    RoleDto getRoleById(Integer id);
+    RoleDto getRoleById(Long id);
     RoleDto createRole(RoleDto request);
-    RoleDto updateRole(Integer id, RoleDto request);
-    void deleteRole(Integer id);
+    RoleDto updateRole(Long id, RoleDto request);
+    void deleteRole(Long id);
 }
