+++ server/auth-service/src/main/java/com/project/authservice/service/impl/RoleServiceImpl.java
@@ -34,7 +34,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional(readOnly = true)
-    public RoleDto getRoleById(Integer id) {
+    public RoleDto getRoleById(Long id) {
         Role role = roleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Role not found"));
         return mapToDto(role);
@@ -65,7 +65,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional
-    public RoleDto updateRole(Integer id, RoleDto request) {
+    public RoleDto updateRole(Long id, RoleDto request) {
         Role role = roleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Role not found"));
                 
@@ -89,7 +89,7 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     @Transactional
-    public void deleteRole(Integer id) {
+    public void deleteRole(Long id) {
         Role role = roleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Role not found"));
         roleRepository.delete(role);
